+++ code/include/FixPolicy.hpp
@@ -5,8 +5,8 @@
  *        It includes exception handling and utility functions for modifying variables.
  * 
  * @author Francesco Biscaccia Carrara
- * @version v1.0.5
- * @since 03/29/2025
+ * @version v1.0.6
+ * @since 04/04/2025
  */