+++ code/include/OMIP.hpp
@@ -8,8 +8,8 @@
  * It inherits core functionality from the base MIP class while adding optimality-specific features.
  * 
  * @author Francesco Biscaccia Carrara
- * @version v1.0.5
- * @since 03/29/2025
+ * @version v1.0.6
+ * @since 04/04/2025
  */
 
 #ifndef OMIP_H