+++ code/include/RlxFMIP.hpp
@@ -9,8 +9,8 @@
  * to restore original problem formulations.
  * 
  * @author Francesco Biscaccia Carrara
- * @version v1.0.5
- * @since 03/29/2025
+ * @version v1.0.6
+ * @since 04/04/2025
  */
 
 #ifndef RlxFMIP_H